Feat: api 에러 수정 #3
