+++ src/components/dictionary/identity/IdentityThumnailList.tsx
@@ -1,17 +1,17 @@
 "use client";
 
 import React, { useState, useEffect, useRef, useCallback } from "react";
-import { Button, Input } from "@material-tailwind/react";
-
+import { Button, Input, Tooltip } from "@material-tailwind/react";
+import { FaCheckCircle, FaRegCircle } from "react-icons/fa";
 import { LuSearch } from "react-icons/lu";
 import axios from "axios";
-import { getEgo } from "@/api/ditionaryApi";
+import { getIdentity } from "@/api/ditionaryApi";
 import useStore from "@/zustand/store"; // zustand 스토어 import
-import EgoThumbnailCard from "./EgoThumbnailCard";
+import IdentityThumbnailCard from "./IdentityThumbnailCard";
 import { Spinner } from "@material-tailwind/react";
 import ErrorMessage from "@/ui/ErrorMessage";
-
-import Filter from "./EgoFilter";
+import nicknamesData from "@/constants/nicknames.json";
+import Filter from "./IdentityFilter";
 import { useQuery } from "@tanstack/react-query";
 
 interface FilterModalProps {
@@ -58,16 +58,20 @@ const FilterModal: React.FC<FilterModalProps> = ({
 };
 
 const TopTitleAndThumnailList = () => {
+  const [isSync, setIsSync] = useState(false);
   const [searchTerm, setSearchTerm] = useState("");
   const [openFilter, setOpenFilter] = useState(false);
   const options = useStore((state) => state.optionsState); // options 상태 가져오기
 
+  const [nicknames, setNicknames] = useState<{ [key: string]: string[] }>({});
+  const [filteredData, setFilteredData] = useState([]);
+  const [paginatedData, setPaginatedData] = useState([]);
   const [page, setPage] = useState(1);
   const observerElem = useRef<HTMLDivElement | null>(null);
 
   const { data, isLoading, isError, error } = useQuery({
-    queryKey: ["ego", options],
-    queryFn: () => getEgo(options),
+    queryKey: ["identity", options],
+    queryFn: () => getIdentity(options),
     staleTime: 1000 * 60 * 60 * 24, // 하루
     refetchOnWindowFocus: false, // 포커스 할 때마다 다시 불러오는 기능 끔
   });
@@ -93,24 +97,55 @@ const TopTitleAndThumnailList = () => {
     };
   }, [handleObserver]);
 
-  const filteredData =
-    data &&
-    data
-      .filter((item: { name: string }) =>
-        item.name.toLowerCase().includes(searchTerm.toLowerCase())
-      )
-      .reverse();
+  useEffect(() => {
+    const nicknamesMap: { [key: string]: string[] } = {};
+    nicknamesData.forEach((item: { id: string; nicknames: string[] }) => {
+      nicknamesMap[item.id] = item.nicknames;
+    });
+    setNicknames(nicknamesMap);
+  }, []);
 
-  const paginatedData = filteredData.slice(0, page * 15);
+  useEffect(() => {
+    const filtered =
+      data &&
+      data
+        .filter((item: { id: string; name: string }) => {
+          const nameMatch = item.name
+            .toLowerCase()
+            .includes(searchTerm.toLowerCase());
+          const nicknameMatch =
+            nicknames[item.id]?.some((nickname) =>
+              nickname.toLowerCase().includes(searchTerm.toLowerCase())
+            ) || false;
+          return nameMatch || nicknameMatch;
+        })
+        .reverse();
+    setFilteredData(filtered || []);
+  }, [data, searchTerm, nicknames]);
+
+  useEffect(() => {
+    if (data) {
+      const filtered = data
+        .filter((item: { name: string }) =>
+          item.name.toLowerCase().includes(searchTerm.toLowerCase())
+        )
+        .reverse();
+      setFilteredData(filtered);
+      setPaginatedData(filtered.slice(0, page * 15));
+    }
+  }, [data, searchTerm, page]);
+
+  const LastestIdentity = filteredData[0].name;
 
   return (
-    <div className="flex-auto md:pl-10">
-      {/* 상단 제목, 버튼 */}
+    <>
       <div className="flex justify-between items-center">
         <span className="text-3xl lg:text-4xl whitespace-nowrap hidden lg:block pr-2">
-          에고
+          인격
         </span>
         <div className="my-2 grid grid-cols-1 sm:flex sm:justify-between w-full lg:w-fit gap-2 h-fit md:h-10">
+          <span>{LastestIdentity}</span>
+
           <Button
             className="h-8 lg:hidden bg-primary-400 lg:h-8 py-0.5 px-4 text-lg lg:text-sm text-primary-100 hover:bg-primary-300 rounded"
             onClick={() => setOpenFilter(true)}
@@ -119,6 +154,25 @@ const TopTitleAndThumnailList = () => {
             <span className="whitespace-nowrap">필터</span>
           </Button>
           <div className="flex gap-2">
+            <Tooltip
+              className="bg-primary-500 text-primary-100 text-xs"
+              content={
+                <span>체크 시 3 동기화 후 이미지를 확인할 수 있습니다.</span>
+              }
+            >
+              <Button
+                className="min-w-[80px] flex gap-2 items-center bg-primary-400 px-2 md:px-4 py-0 md:py-1 font-sansLight text-sm md:text-base text-white hover:bg-primary-300 rounded"
+                placeholder={undefined}
+                onClick={() => setIsSync((prev) => !prev)}
+              >
+                <span className="pt-1 whitespace-nowrap">동기화</span>
+                {isSync ? (
+                  <FaCheckCircle className="text-primary-200" />
+                ) : (
+                  <FaRegCircle className="text-primary-200" />
+                )}
+              </Button>
+            </Tooltip>
             <div className="relative flex w-full gap-2 md:w-max">
               <Input
                 type="search"
@@ -142,10 +196,7 @@ const TopTitleAndThumnailList = () => {
           </div>
         </div>
       </div>
-
       <FilterModal openFilter={openFilter} setOpenFilter={setOpenFilter} />
-
-      {/* 썸네일 리스트 */}
       {isLoading ? (
         <div className="flex justify-center items-center h-screen">
           <Spinner className="w-8 h-8 text-primary-200" />
@@ -153,7 +204,7 @@ const TopTitleAndThumnailList = () => {
       ) : isError ? (
         axios.isAxiosError(error) && error.response?.status === 404 ? (
           <div className="text-primary-200 text-center w-full my-8 h-screen">
-            해당하는 에고가 없습니다.
+            해당하는 인격이 없습니다.
           </div>
         ) : (
           <div className="text-primary-200 text-center w-full my-8">
@@ -170,19 +221,22 @@ const TopTitleAndThumnailList = () => {
                   name: string;
                   grade: number;
                   character: string;
-                  zoomImage: string;
-                  image: string;
+                  beforeImage: string | null;
+                  afterImage: string | null;
                 },
                 index: number
               ) => (
-                <EgoThumbnailCard
+                <IdentityThumbnailCard
                   key={index}
                   id={item.id}
                   grade={item.grade}
                   name={item.name}
                   character={item.character}
-                  imageZoomIn={item.zoomImage}
-                  imageZoomOut={item.image}
+                  imageBefore={
+                    item.beforeImage ? item.beforeImage.trimEnd() : ""
+                  }
+                  imageAfter={item.afterImage ? item.afterImage.trimEnd() : ""}
+                  isSync={isSync}
                 />
               )
             )
@@ -194,7 +248,7 @@ const TopTitleAndThumnailList = () => {
         </div>
       )}
       <div ref={observerElem} className="h-10"></div>
-    </div>
+    </>
   );
 };
